rearrange(server): create top-tracks controllers for top-track routes
